+++ render/render.go
@@ -1,9 +1,9 @@
 package render
 
 import (
+	"bytes"
 	"slices"
 
-	"github.com/PiquelOrganization/docs.piquel.fr/source"
 	"github.com/gomarkdown/markdown"
 	"github.com/gomarkdown/markdown/ast"
 	"github.com/gomarkdown/markdown/html"
@@ -18,58 +18,6 @@ const tailwindBase = `
     h5 { font-size: 0.83em; }
     h6 { font-size: 0.67em; }
 `
-
-type Renderer interface {
-	RenderAllFiles(config *RenderConfig) (map[string][]byte, error)
-	RenderFile(path string, config *RenderConfig) ([]byte, error)
-}
-
-type RenderConfig struct {
-	RootPath    string // this will be prepended to any local URLs in the markdown
-	UseTailwind bool   // wether to use tailwind classes and settings (notably restore the proper size of titles)
-	FullPage    bool   // wether to render a full page (add <!DOCTYPE html> to the top of the page
-}
-
-type RealRenderer struct {
-	source source.Source
-}
-
-func NewRealRenderer(source source.Source) Renderer {
-	return &RealRenderer{source: source}
-}
-
-func (r *RealRenderer) RenderAllFiles(config *RenderConfig) (map[string][]byte, error) {
-	files := map[string][]byte{}
-	fileNames, err := r.source.GetAllMarkdown()
-	if err != nil {
-		return map[string][]byte{}, err
-	}
-
-	for _, file := range fileNames {
-		renderedFile, err := r.RenderFile(file, config)
-		if err != nil {
-			return map[string][]byte{}, err
-		}
-		files[file] = renderedFile
-	}
-	return files, nil
-}
-
-func (r *RealRenderer) RenderFile(path string, config *RenderConfig) ([]byte, error) {
-	file, err := r.source.LoadFile(path)
-	if err != nil {
-		return []byte{}, err
-	}
-
-	custom, err := r.renderCustom(file, config)
-	if err != nil {
-		return []byte{}, err
-	}
-	doc := r.parseMarkdown(custom, config)
-	doc = r.fixupAST(doc, config)
-	html := r.renderHTML(doc, config)
-	return r.addStyles(html, config), nil
-}
 
 func (r *RealRenderer) parseMarkdown(md []byte, config *RenderConfig) ast.Node {
 	extensions := parser.CommonExtensions | parser.AutoHeadingIDs | parser.NoEmptyLineBeforeBlock
@@ -103,4 +51,28 @@ func (r *RealRenderer) addStyles(html []byte, config *RenderConfig) []byte {
 
 	styles = slices.Concat([]byte("<style>\n"), styles, []byte("</style>\n"))
 	return slices.Concat(html, styles)
+}
+
+func (r *RealRenderer) fixupAST(doc ast.Node, config *RenderConfig) ast.Node {
+	ast.WalkFunc(doc, func(node ast.Node, entering bool) ast.WalkStatus {
+		switch node := node.(type) {
+		case *ast.Link:
+			r.fixupLink(node, entering, config)
+		}
+
+		return ast.GoToNext
+	})
+	return doc
+}
+
+func (r *RealRenderer) fixupLink(link *ast.Link, entering bool, config *RenderConfig) {
+	if !entering {
+		return
+	}
+
+	if bytes.HasPrefix(link.Destination, []byte("http")) {
+		link.AdditionalAttributes = append(link.AdditionalAttributes, "target=\"_blank\"")
+	} else {
+		link.Destination = slices.Concat([]byte(config.RootPath), utils.FormatLocalPath(link.Destination, ".md"))
+	}
 }
